pyproject.toml
# This file is part of REANA.
# Copyright (C) 2025 CERN.
#
# REANA is free software; you can redistribute it and/or modify it
# under the terms of the MIT License; see LICENSE file for more details.

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "reana-dask-kubernetes-operator"
dynamic = ["version"]
authors = [
    { name = "REANA", email = "info@reana.io" }
]
description = "REANA-Dask-Kubernetes-Operator"
license = {file = "LICENSE"}
readme = "README.md"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Web Environment",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: Implementation :: CPython",
    "Programming Language :: Python",
    "Topic :: Internet :: WWW/HTTP :: Dynamic Content",
    "Topic :: Software Development :: Libraries :: Python Modules"
]
requires-python = ">=3.8"
dependencies = [
    "reana-commons==0.95.0a9",
    "reana-db==0.95.0a5"
]

[project.optional-dependencies]
test = [
    "pytest-reana>=0.95.0a4,<0.96.0",
]

[project.urls]
Homepage = "https://github.com/reanahub/reana-dask-kubernetes-operator"

[project.entry-points."dask_operator_plugin"]
reana_dask_kubernetes_operator = "reana_dask_kubernetes_operator.watcher"

[tool.setuptools.packages.find]
include = ["reana_dask_kubernetes_operator"]

[tool.setuptools.dynamic]
version = {attr = "reana_dask_kubernetes_operator.version.__version__"}
